refactor(flow_designer): split 1328-LOC module into a package
Split src/servicenow_mcp/tools/flow_designer.py into:
- _definition.py: flow_list, flow_get, flow_map, flow_snapshot_list
- _action.py: flow_action_detail (two-phase fetch preserved)
- _execution.py: flow_execution_list, flow_execution_detail
- _migration.py: workflow_migration_analysis + all migration helpers
- __init__.py: coordinator; re-exports register_tools, TOOL_NAMES, _process_neighbor

Public API and registration order preserved. tests/test_flow_designer.py
imports _process_neighbor unchanged. 1202 tests pass.

Closes audit refactor R2.
