+++ src/servicenow_mcp/tools/flow_designer/_migration.py
@@ -1,9 +1,8 @@
-"""Flow Designer introspection and migration analysis tools."""
+"""Legacy workflow to Flow Designer migration analysis tool and helpers."""
 
 from __future__ import annotations
 
 import asyncio
-import logging
 import re
 from typing import Any
 
@@ -26,8 +25,6 @@ from servicenow_mcp.utils import (
     validate_identifier,
 )
 
-
-logger = logging.getLogger(__name__)
 
 # -- Duplicated string constants ---------------------------------------------------
 
@@ -843,449 +840,8 @@ def _assemble_migration_response(
     }
 
 
-def _resolve_flow_map_target(flow_record: dict[str, Any], original_flow_sys_id: str) -> str:
-    """Resolve the child-link target used by Flow Designer map tables."""
-    latest_snapshot_sys_id = resolve_ref_value(flow_record.get("latest_snapshot", ""))
-    if latest_snapshot_sys_id:
-        return latest_snapshot_sys_id
-
-    master_snapshot_sys_id = resolve_ref_value(flow_record.get("master_snapshot", ""))
-    if master_snapshot_sys_id:
-        return master_snapshot_sys_id
-
-    return original_flow_sys_id
-
-
-TOOL_NAMES: list[str] = [
-    "flow_list",
-    "flow_get",
-    "flow_map",
-    "flow_action_detail",
-    "flow_execution_list",
-    "flow_execution_detail",
-    "flow_snapshot_list",
-    "workflow_migration_analysis",
-]
-
-
-def register_tools(mcp: FastMCP, settings: Settings, auth_provider: BasicAuthProvider) -> None:
-    """Register Flow Designer introspection and migration tools on the MCP server."""
-
-    @mcp.tool()
-    @tool_handler
-    async def flow_list(
-        table: str = "",
-        flow_type: str = "",
-        status: str = "",
-        active_only: bool = True,
-        limit: int = 20,
-        *,
-        correlation_id: str = "",
-    ) -> str:
-        """List Flow Designer flows and subflows with optional filters.
-
-        Args:
-            table: Filter by the table the flow operates on (e.g. 'incident').
-            flow_type: Filter by flow type (e.g. 'flow', 'subflow', 'action').
-            status: Filter by flow status (e.g. 'draft', 'published').
-            active_only: Only return active flows (default True).
-            limit: Maximum number of flows to return (default 20).
-        """
-        if table:
-            validate_identifier(table)
-        check_table_access("sys_hub_flow")
-
-        query = (
-            ServiceNowQuery()
-            .equals_if("table", table, bool(table))
-            .equals_if("type", flow_type, bool(flow_type))
-            .equals_if("status", status, bool(status))
-            .equals_if("active", "true", active_only)
-            .build()
-        )
-        safety = enforce_query_safety("sys_hub_flow", query, limit, settings)
-
-        async with ServiceNowClient(settings, auth_provider) as client:
-            result = await client.query_records(
-                "sys_hub_flow",
-                query,
-                fields=[
-                    "sys_id",
-                    "name",
-                    "status",
-                    "type",
-                    "table",
-                    "active",
-                    "description",
-                    "sys_updated_on",
-                    "sys_created_on",
-                ],
-                limit=safety["limit"],
-                display_values=True,
-            )
-
-        flows = [mask_sensitive_fields(f) for f in result["records"]]
-
-        return format_response(
-            data={"flows": flows},
-            correlation_id=correlation_id,
-        )
-
-    @mcp.tool()
-    @tool_handler
-    async def flow_get(
-        flow_sys_id: str,
-        *,
-        correlation_id: str = "",
-    ) -> str:
-        """Fetch a Flow Designer flow definition with its input and output variables.
-
-        Args:
-            flow_sys_id: The sys_id of the sys_hub_flow record.
-        """
-        validate_identifier(flow_sys_id)
-        check_table_access("sys_hub_flow")
-        check_table_access("sys_hub_flow_variable")
-
-        var_query = ServiceNowQuery().equals("flow", flow_sys_id).build()
-        var_safety = enforce_query_safety("sys_hub_flow_variable", var_query, 100, settings)
-
-        async with ServiceNowClient(settings, auth_provider) as client:
-            flow_record, var_result = await asyncio.gather(
-                client.get_record("sys_hub_flow", flow_sys_id, display_values=True),
-                client.query_records(
-                    "sys_hub_flow_variable",
-                    var_query,
-                    fields=[
-                        "sys_id",
-                        "name",
-                        "type",
-                        "mandatory",
-                        "default_value",
-                    ],
-                    limit=var_safety["limit"],
-                    display_values=True,
-                ),
-            )
-
-        return format_response(
-            data={
-                "flow": mask_sensitive_fields(flow_record),
-                "variables": [mask_sensitive_fields(v) for v in var_result["records"]],
-            },
-            correlation_id=correlation_id,
-        )
-
-    @mcp.tool()
-    @tool_handler
-    async def flow_map(
-        flow_sys_id: str,
-        *,
-        correlation_id: str = "",
-    ) -> str:
-        """Map the structure of a Flow Designer flow: action instances and logic blocks.
-
-        Args:
-            flow_sys_id: The sys_id of the sys_hub_flow record.
-        """
-        validate_identifier(flow_sys_id)
-        check_table_access("sys_hub_flow")
-        check_table_access("sys_hub_action_instance")
-        check_table_access("sys_hub_flow_logic")
-
-        async with ServiceNowClient(settings, auth_provider) as client:
-            # display_values=False is intentional so latest_snapshot/master_snapshot stay raw sys_ids.
-            flow_record = await client.get_record("sys_hub_flow", flow_sys_id, display_values=False)
-            flow_map_target = _resolve_flow_map_target(flow_record, flow_sys_id)
-
-            action_query = ServiceNowQuery().equals("flow", flow_map_target).order_by("position").build()
-            logic_query = ServiceNowQuery().equals("flow", flow_map_target).order_by("position").build()
-
-            action_safety = enforce_query_safety("sys_hub_action_instance", action_query, 100, settings)
-            logic_safety = enforce_query_safety("sys_hub_flow_logic", logic_query, 100, settings)
-
-            action_result, logic_result = await asyncio.gather(
-                client.query_records(
-                    "sys_hub_action_instance",
-                    action_query,
-                    fields=[
-                        "sys_id",
-                        "name",
-                        "action_type",
-                        "order",
-                        "position",
-                        "sys_created_on",
-                    ],
-                    limit=action_safety["limit"],
-                    display_values=True,
-                ),
-                client.query_records(
-                    "sys_hub_flow_logic",
-                    logic_query,
-                    fields=[
-                        "sys_id",
-                        "name",
-                        "type",
-                        "order",
-                        "position",
-                        "sys_created_on",
-                    ],
-                    limit=logic_safety["limit"],
-                    display_values=True,
-                ),
-            )
-
-        return format_response(
-            data={
-                "actions": [mask_sensitive_fields(a) for a in action_result["records"]],
-                "logic_blocks": [mask_sensitive_fields(lb) for lb in logic_result["records"]],
-            },
-            correlation_id=correlation_id,
-        )
-
-    @mcp.tool()
-    @tool_handler
-    async def flow_action_detail(
-        action_instance_sys_id: str,
-        *,
-        correlation_id: str = "",
-    ) -> str:
-        """Fetch detailed information about a flow action instance, its type definition, and steps.
-
-        Args:
-            action_instance_sys_id: The sys_id of the sys_hub_action_instance record.
-        """
-        validate_identifier(action_instance_sys_id)
-        check_table_access("sys_hub_action_instance")
-        check_table_access("sys_hub_action_type_definition")
-        check_table_access("sys_hub_step_instance")
-
-        async with ServiceNowClient(settings, auth_provider) as client:
-            # Phase 1: raw instance to extract the action_type sys_id
-            raw_instance = await client.get_record(
-                "sys_hub_action_instance",
-                action_instance_sys_id,
-                display_values=False,
-            )
-            action_type_sys_id = resolve_ref_value(raw_instance.get("action_type", ""))
-
-            # Phase 2: parallel fetch of display instance, type definition (if linked), and steps
-            if not action_type_sys_id:
-                display_instance = await client.get_record(
-                    "sys_hub_action_instance",
-                    action_instance_sys_id,
-                    display_values=True,
-                )
-                type_definition = None
-                step_records: list[dict[str, Any]] = []
-            else:
-                validate_identifier(action_type_sys_id)
-
-                steps_query = ServiceNowQuery().equals("action", action_type_sys_id).order_by("order").build()
-                steps_safety = enforce_query_safety("sys_hub_step_instance", steps_query, 100, settings)
-
-                display_instance, type_definition, steps_result = await asyncio.gather(
-                    client.get_record(
-                        "sys_hub_action_instance",
-                        action_instance_sys_id,
-                        display_values=True,
-                    ),
-                    client.get_record(
-                        "sys_hub_action_type_definition",
-                        action_type_sys_id,
-                        display_values=True,
-                    ),
-                    client.query_records(
-                        "sys_hub_step_instance",
-                        steps_query,
-                        fields=[
-                            "sys_id",
-                            "name",
-                            "step_type",
-                            "order",
-                            "sys_created_on",
-                        ],
-                        limit=steps_safety["limit"],
-                        display_values=True,
-                    ),
-                )
-                step_records = steps_result["records"]
-
-        return format_response(
-            data={
-                "instance": mask_sensitive_fields(display_instance),
-                "type_definition": mask_sensitive_fields(type_definition) if type_definition else None,
-                "steps": [mask_sensitive_fields(s) for s in step_records],
-            },
-            correlation_id=correlation_id,
-        )
-
-    @mcp.tool()
-    @tool_handler
-    async def flow_execution_list(
-        flow_sys_id: str = "",
-        source_record: str = "",
-        state: str = "",
-        limit: int = 20,
-        *,
-        correlation_id: str = "",
-    ) -> str:
-        """List Flow Designer execution contexts with optional filters.
-
-        Args:
-            flow_sys_id: Filter by the sys_id of the flow definition.
-            source_record: Filter by the sys_id of the source record that triggered the flow.
-            state: Filter by execution state (e.g. 'IN_PROGRESS', 'COMPLETE', 'ERROR', 'CANCELLED').
-            limit: Maximum number of execution contexts to return (default 20).
-        """
-        if not flow_sys_id and not source_record:
-            return format_response(
-                data=None,
-                correlation_id=correlation_id,
-                status="error",
-                error="At least one of flow_sys_id or source_record is required",
-            )
-
-        if flow_sys_id:
-            validate_identifier(flow_sys_id)
-        if source_record:
-            validate_identifier(source_record)
-        check_table_access("sys_flow_context")
-
-        query = (
-            ServiceNowQuery()
-            .equals_if("flow", flow_sys_id, bool(flow_sys_id))
-            .equals_if("source_record", source_record, bool(source_record))
-            .equals_if("state", state, bool(state))
-            .build()
-        )
-        safety = enforce_query_safety("sys_flow_context", query, limit, settings)
-
-        async with ServiceNowClient(settings, auth_provider) as client:
-            result = await client.query_records(
-                "sys_flow_context",
-                query,
-                fields=[
-                    "sys_id",
-                    "name",
-                    "flow",
-                    "source_record",
-                    "source_table",
-                    "state",
-                    "started",
-                    "ended",
-                    "sys_created_on",
-                ],
-                limit=safety["limit"],
-                display_values=True,
-            )
-
-        executions = [mask_sensitive_fields(e) for e in result["records"]]
-
-        return format_response(
-            data={"executions": executions},
-            correlation_id=correlation_id,
-        )
-
-    @mcp.tool()
-    @tool_handler
-    async def flow_execution_detail(
-        context_id: str,
-        *,
-        correlation_id: str = "",
-    ) -> str:
-        """Fetch detailed execution information for a Flow Designer context including ordered log entries.
-
-        Provides richer detail than debug_flow_execution, including operation types, log levels,
-        and duration per step.
-
-        Args:
-            context_id: The sys_id of the sys_flow_context record.
-        """
-        validate_identifier(context_id)
-        check_table_access("sys_flow_context")
-        check_table_access("sys_flow_log")
-
-        log_query = ServiceNowQuery().equals("context", context_id).order_by("order").build()
-        log_safety = enforce_query_safety("sys_flow_log", log_query, 500, settings)
-
-        async with ServiceNowClient(settings, auth_provider) as client:
-            context_record, log_result = await asyncio.gather(
-                client.get_record("sys_flow_context", context_id, display_values=True),
-                client.query_records(
-                    "sys_flow_log",
-                    log_query,
-                    fields=[
-                        "sys_id",
-                        "action",
-                        "operation",
-                        "level",
-                        "message",
-                        "order",
-                        "sys_created_on",
-                        "output_data",
-                        "error_message",
-                        "duration",
-                    ],
-                    limit=log_safety["limit"],
-                    display_values=True,
-                ),
-            )
-
-        logs = [mask_sensitive_fields(entry) for entry in log_result["records"]]
-
-        return format_response(
-            data={
-                "context": mask_sensitive_fields(context_record),
-                "log_count": len(logs),
-                "logs": logs,
-            },
-            correlation_id=correlation_id,
-        )
-
-    @mcp.tool()
-    @tool_handler
-    async def flow_snapshot_list(
-        flow_sys_id: str,
-        limit: int = 20,
-        *,
-        correlation_id: str = "",
-    ) -> str:
-        """List published snapshots (versions) for a Flow Designer flow.
-
-        Args:
-            flow_sys_id: The sys_id of the sys_hub_flow record.
-            limit: Maximum number of snapshots to return (default 20).
-        """
-        validate_identifier(flow_sys_id)
-        check_table_access("sys_hub_flow_snapshot")
-
-        query = ServiceNowQuery().equals("parent_flow", flow_sys_id).order_by("sys_created_on", descending=True).build()
-        safety = enforce_query_safety("sys_hub_flow_snapshot", query, limit, settings)
-
-        async with ServiceNowClient(settings, auth_provider) as client:
-            result = await client.query_records(
-                "sys_hub_flow_snapshot",
-                query,
-                fields=[
-                    "sys_id",
-                    "name",
-                    "parent_flow",
-                    "version",
-                    "status",
-                    "sys_created_on",
-                    "sys_updated_on",
-                ],
-                limit=safety["limit"],
-                display_values=True,
-            )
-
-        snapshots = [mask_sensitive_fields(s) for s in result["records"]]
-
-        return format_response(
-            data={"snapshots": snapshots},
-            correlation_id=correlation_id,
-        )
+def register(mcp: FastMCP, settings: Settings, auth_provider: BasicAuthProvider) -> None:
+    """Register the workflow_migration_analysis tool."""
 
     @mcp.tool()
     @tool_handler
